garden_chatbot3/requirements.txt
Flask
google-generativeai
gunicorn
gTTS